requirements.txt
pandas
numpy
requests
nltk
vaderSentiment
textblob
scikit-learn
python-dotenv
transformers
torch
matplotlib
groq